// src/modules/lead/enums/lead.enum.ts
export enum ENUM_LEAD_STATUS {
    NEW = 'new',
    CONTACTED = 'contacted',
    QUALIFIED = 'qualified',
    PROPOSAL_SENT = 'proposal_sent',
    WON = 'won',
    LOST = 'lost',
}

export enum ENUM_LEAD_SOURCE {
    WEB = 'web',
    REFERRAL = 'referral',
    TRADE_SHOW = 'trade_show',
    COLD_CALL = 'cold_call',
    EXISTING_CUSTOMER = 'existing_customer',
    OTHER = 'other',
}